+++ docker/setenv.sh
@@ -1,0 +1,8 @@
+# External URL where the MXM service will be accessible:
+export MXM_EXTERNAL_URL="http://localhost:8080"
+
+# (required) Password for the database "mxm" user:
+export POSTGRES_PASSWORD=
+
+# Host directory for the postgres database:
+export POSTGRES_DATA="$(pwd)/db_data"