+++ ChangeLog.md
@@ -1,4 +1,5 @@
 2022-08
 
+- docker build; no history mode yet; other adjustments
 - spa routing and docker preps
 - initial (re)implementation of the MXM Server component